+++ pom.xml
@@ -51,7 +51,7 @@
         <module>services/xwiki-spring-boot-starter</module>
         <module>services/image-processing</module>
         <module>services/icecat</module>
-
+        <module>services/blog</module>
