+++ ui/src/main/java/org/open4goods/ui/controllers/ui/VerticalBrandsController.java
@@ -13,7 +13,7 @@ import org.open4goods.model.vertical.SubsetCriteriaOperator;
 import org.open4goods.model.vertical.VerticalConfig;
 import org.open4goods.model.vertical.VerticalSubset;
 import org.open4goods.services.serialisation.service.SerialisationService;
-import org.open4goods.ui.services.BlogService;
+import org.open4goods.services.blog.service.BlogService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.web.servlet.ModelAndView;
