+++ ui/src/main/java/org/open4goods/ui/controllers/ui/pages/BlogController.java
@@ -5,8 +5,8 @@ import java.util.List;
 
 import org.open4goods.ui.config.yml.UiConfig;
 import org.open4goods.ui.controllers.ui.UiService;
-import org.open4goods.ui.model.BlogPost;
-import org.open4goods.ui.services.BlogService;
+import org.open4goods.services.blog.model.BlogPost;
+import org.open4goods.services.blog.service.BlogService;
 import org.open4goods.xwiki.services.XWikiHtmlService;
 import org.open4goods.xwiki.services.XwikiFacadeService;
 import org.slf4j.Logger;
